Moved the view refresher to the events package
